refactor: plugin descriptor finder
